requirements.txt
streamlit
langchain
pydantic
python-dotenv
PyPDF2
faiss-cpu
sentence-transformers
huggingface_hub
langchain-community
langchain-huggingface
nest-asyncio